tests/wpt/tests/pointerlock/pointerlock_leave_Tab-manual.html

<!DOCTYPE html>
<html>
<body>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name='flags' content='interact'>
<style type="text/css">
    button {
        color: blue;
    }

    #target-wrap {
        position: relative;
        background-color: lightgrey;
        width: 200px;
        height: 100px;
        border: grey 1px solid;
    }

    #target {
        position: relative;
        background-color: lightyellow;
        width: 100px;
        height: 30px;
        border: yellow 1px solid;
    }

    #status-log {
        margin: 10px 0;
        color: green;
    }
</style>
</head>
<body>
    <h2>Description</h2>
    <p>This test validates that pointer lock will be lost the user agent / window loses focus.</p>
    <hr/>

    <h2>Manual Test Steps:</h2>
    <p>
        <ol>
            <li>Click the "lockTarget" button to request a pointer lock.</li>
            <li>Focus to another tab with keyboard (Ctrl-TAB).</li>
            <li>Test is done.</li>
        </ol>
    </p>
    <hr/>

    <button onclick="lockTarget();">lockTarget</button>

    <div id="target-wrap">
        <div id="status-log">Click the "lockTarget" button.</div>
        <div id="target">Target</div>
    </div>
    <hr/>

    <div id="log"></div>

    <script type="text/javascript" >
        var target = document.querySelector('#target'),
            status_log = document.querySelector('#status-log');

        var leaveTabTest = async_test("Test that pointer lock will be lost when the current Tab loses focus.");

    document.addEventListener("pointerlockchange", function() {
        if(document.pointerLockElement) {
            status_log.innerHTML = "Please leave the current tab.";
        } else {
            status_log.innerHTML = "Pointer lock exited!";

            leaveTabTest.step(function() {
                assert_equals(document.pointerLockElement, null, "Pointer lock exited!");
            });

            leaveTabTest.done();
        }
    });

    function lockTarget() {
        target.requestPointerLock();
    }
    </script>
    </body>
</html>
